upd: separate test bootstrap file
